Created Consumption History Bar Chart. Merged History bar chart and success pie chart into a single element
